feat (layout): Refactorizar sidebar, header y /app/(protected)/layout.tsx
- Refactorización de la estructura del sidebar para organización y reutilización
- Renombrar DashboardHeader como MobileHeader
- Refactorización del MobileHeader para integrar la nueva sidebar
- Refactorización del /app/(protected)/layout.tsx para mostrar el MobileHeader solo en mobile
