+++ app/(protected)/layout.tsx
@@ -1,11 +1,10 @@
 import type React from "react"
 import type { Metadata } from "next/types"
 import { Inter } from "next/font/google"
-import "../globals.css"
-import { AppSidebar } from "@/components/sidebar"
-import { DashboardHeader } from "@/components/dashboard-header"
-import { SidebarProvider } from "@/components/ui/use-sidebar" // Importa SidebarProvider
-import { getFullUserSession } from "@/lib/auth" // Importa getFullUserSession
+import { AppSidebar } from "@/components/sidebar/sidebar"
+import { MobileHeader } from "@/components//header/mobile-header"
+import { SidebarProvider } from "@/components/ui/use-sidebar"
+import { getFullUserSession } from "@/lib/auth"
 import { ThemeProvider } from "@/components/theme-provider"
 
 const inter = Inter({ subsets: ["latin"] })
@@ -20,20 +19,21 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  const session = await getFullUserSession() // Obtiene la sesión completa del usuario
+  const session = await getFullUserSession()
 
   return (
     <html lang="es" suppressHydrationWarning>
       <body className={inter.className}>
         {/* <ThemeProvider attribute="class" defaultTheme="light" enableSystem disableTransitionOnChange> */}
         <SidebarProvider>
-          <AppSidebar userRole={session?.role || null} userEmail={session?.email || null} hide={false} />
+          <AppSidebar session={session} />
           <div
             className="flex flex-1 flex-col min-h-svh bg-background w-full overflow-x-hidden
-                         md:data-[state=collapsed]:group-[.peer]:ml-(--sidebar-width-icon)
+                         md:group-[.peer]:data-[state=collapsed]:ml-[var(--sidebar-width-icon)]
                          transition-[margin-left] duration-200 ease-linear"
           >
-            <DashboardHeader />
+            <MobileHeader session={session} />
+
             <main className="flex flex-1 flex-col gap-4 p-4 md:gap-8 md:p-6 min-w-0 max-w-full">{children}</main>
           </div>
         </SidebarProvider>
@@ -42,4 +42,3 @@ export default async function RootLayout({
     </html>
   )
 }
-
