+++ components/nav-interface.ts
@@ -1,0 +1,13 @@
+interface NavItem {
+    title: string
+    href: string
+    roles: string[]
+}
+
+export interface NavSection {
+    title: string
+    items: NavItem[]
+    icon: React.ElementType
+    expandable?: boolean
+    isActive?: boolean
+}
